修复合成确认界面侧边栏、样板配方类型和 AE2Proxy 反射
- CraftConfirmScreenMixin: call render() and drawForeground() so sidebar
  hover effects, tooltips, and favorites work correctly
- AE2Proxy: fix getStackUnderMouse reflection using double.class instead of
  int.class to match actual method signature
- ProviderSearchHelper: add class-name fallback for recipe type search key
  when EAEP has no mapping (e.g. ReactionChamberRecipe → "reaction chamber")
- EmiEncodePatternHandlerMixin: improve debug logging for pattern encoding

+++ src/main/java/org/chatterjay/emiextend/integration/AE2Proxy.java
@@ -54,7 +54,7 @@ public class AE2Proxy {
             Class<?> ccsClass = Class.forName("appeng.client.gui.me.crafting.CraftConfirmScreen");
             if (!ccsClass.isInstance(screen)) return ItemStack.EMPTY;
 
-            Method getStackMethod = ccsClass.getMethod("getStackUnderMouse", int.class, int.class);
+            Method getStackMethod = ccsClass.getMethod("getStackUnderMouse", double.class, double.class);
             Object swb = getStackMethod.invoke(screen, mouseX, mouseY);
             if (swb == null) return ItemStack.EMPTY;
